Revert incorrect RCL_3 drop: Revision: 201033 Kit: 201035
--HG--
branch : RCL_3

+++ serviceproviders/sapi_logging/tsrc/dev/tloggingservice/src/tdeletetest1.cpp
@@ -87,10 +87,14 @@ TInt DeleteEventsync(void)
 	iter = LogService->GetListL(filter) ;
 	
 	
-	while (aEvent = iter->NextL())
-	{
+	while (1){
+	   aEvent = iter->NextL();
+	   if(aEvent){
 	    LogService->DeleteEventL(aEvent->Id()) ;
 	    delete aEvent ;
+	   }
+	   else
+	      break;
 	}
 	
 	delete iter ;
